pyproject.toml
[project]
name = "pyservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "nas", email = "nas@denk-solutions.com" }
]
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "cachecontrol>=0.14.3",
    "fastapi[standard]>=0.115.12",
    "google-auth>=2.39.0",
    "greenlet>=3.2.1",
    "passlib>=1.7.4",
    "pendulum>=3.1.0",
    "pydantic>=2.11.4",
    "pydantic-extra-types>=2.10.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
]

[project.scripts]
pyservice = "pyservice:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
]
